requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.11
distlib==0.3.4
filelock==3.4.2
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
platformdirs==2.4.1
six==1.16.0
slackclient==2.9.3
virtualenv==20.13.0
yarl==1.7.2
